+++ inplaceeditform_bootstrap/static/js/jquery.inplaceeditform.hooks.js
@@ -1,0 +1,19 @@
+(function ($) {
+    "use strict";
+    var enableActionInplaceEditOriginal = $.inplaceeditform.methods.enableInplaceEditAction;
+    var disableActionInplaceEditOriginal = $.inplaceeditform.methods.disableInplaceEditAction;
+    $.inplaceeditform.extend({
+        enableInplaceEditAction: function(inplace_span) {
+            var self = $.inplaceeditform;
+            enableActionInplaceEditOriginal();
+            self.inplaceeditfields.attr("data-toggle", "tooltip");
+            self.inplaceeditfields.attr("title","Double Click to Edit");
+        },
+        disableInplaceEditAction: function(inplace_span) {
+            var self = $.inplaceeditform;
+            disableActionInplaceEditOriginal();
+            self.inplaceeditfields.removeAttr("data-toggle");
+            self.inplaceeditfields.removeAttr("title");
+        },
+    });
+})(jQuery);