+++ CHANGES.rst
@@ -4,6 +4,7 @@ Releases
 0.1.2 (2013-09-??)
 ------------------
 
+* Add Tooltip on Mouseover
 * Improvements in the testing project
 
 0.1.1 (2013-09-17)